+++ UAM Debug.launch
@@ -82,5 +82,5 @@
     </listAttribute>
     <stringAttribute key="org.eclipse.dsf.launch.MEMORY_BLOCKS" value="&lt;?xml version=&quot;1.0&quot; encoding=&quot;UTF-8&quot; standalone=&quot;no&quot;?&gt;&lt;memoryBlockExpressionList context=&quot;reserved-for-future-use&quot;/&gt;"/>
     <stringAttribute key="process_factory_id" value="com.st.stm32cube.ide.mcu.debug.launch.HardwareDebugProcessFactory"/>
-    <stringAttribute key="saved_expressions&lt;seperator&gt;Unknown" value="0x08060000,08060000,0x8060000,8060120,0x8060120,38800000,0x38800000,0x5200201C,0x52002040,0x5802480C"/>
+    <stringAttribute key="saved_expressions&lt;seperator&gt;Unknown" value="0x08060000,08060000,0x8060000,8060120,0x8060120,38800000,0x38800000,0x5200201C,0x52002040,0x5802480C,0x40040008"/>
 </launchConfiguration>
